# Make call to `Sprintf` lazy