This results in a measurable performance improvement (~ 10% better) on a large test suite due to the number of iterators created

CLA signed and submitted