@barakmich 2016-02-23
```
benchmark                                   old ns/op     new ns/op     delta	mult
BenchmarkNamePredicate                      400559        382713        -4.46%	-1.05x
BenchmarkLargeSetsNoIntersection            6462063       5786903       -10.45%	-1.12x
BenchmarkVeryLargeSetsSmallIntersection     12796114      10848564      -15.22%	-1.18x
BenchmarkHelplessContainsChecker            42974392      38127931      -11.28%	-1.13x
BenchmarkHelplessNotContainsFilms           42733549      38866808      -9.05%	-1.10x
BenchmarkHelplessNotContainsActors          41440420      38027936      -8.23%	-1.09x
BenchmarkNetAndSpeed                        2993519       2864484       -4.31%	-1.05x
BenchmarkKeanuAndNet                        2742137       2617562       -4.54%	-1.05x
BenchmarkKeanuAndSpeed                      2898534       4645030       +60.25%	+1.60x
BenchmarkKeanuOther                         6235815       5927680       -4.94%	-1.05x
BenchmarkKeanuBullockOther                  9494207       8492887       -10.55%	-1.12x
BenchmarkSaveBogartPerformances             42541420      10034930      -76.41%	-4.24x
```

I'll point out the "big win" and "big loss" are due to my laptop hanging on one or the other. Net over net, it seems to be a solid 5-10%-er for the memstore, logically enough, more on deeper queries.